Fix matching parameter and parameter use. (#218)
Description: the generic AST has the notion of an identifier (a string)
and a name (a possibly qualified string). A parameter is always an identifier,
but when it is used, it is converted in a name, which was preventing
a equal match.
This diff fixes that by adding the parameter as a Name in the environment,
so it can be later matched against its use.

This will fix issue #201

Test plan: test file included.