Flexible and correct import matching
This fixes the import matching improvements introduced by Drew and Isaac
by doing the flexible matching only if the pattern is simple.
If the pattern is complex, for example searching for the use of
'import *' in Js or Python, it should not match any kind of import.

This fixes issue #197

Test plan:
Many tests included.