Big step towards user-defined equivalences.
In this diff, we apply a list of equivalences to the sgrep
pattern by visiting the pattern, looking for a match, and substituting
the matched pattern with the right hand part of the equivalence.

This helps fix issue #329

Test plan:
home/pad/github/sgrep/sgrep/_build/default/bin/main_sgrep.exe -lang py -equivalences data/basic_equivalences.yml -e 1+2 tests/EQUIV/
/home/pad/github/sgrep/sgrep/tests/EQUIV/test.py:5
   return 2+1

The next diff needs to add DisjExpr to AST generic so we look
for the original sgrep pattern Or the equivalent one.