Add -macros and -error_recovery flags (#469)
Description: For C (and C++) code, the parser in pfff use some heuristics
to parse C code as is, before cpp. However, sometimes certain constructs,
e.g. foreach_list() {} lead to parse error and are not currently handled.
To extend those heuristic one can
pass a -macros argument that will contain a file with macros to
expand (here the definition or a fake definition of foreach_list).

This partially solves issue #448

Test plan:
$ cat ~/work/lang-c/git/pfff_macros.h

$ sgrep -e '$X++' -macros ~/work/lang-c/git/pfff_macros.h -lang c -error_recovery ~/work/lang-c/git/ 2>/dev/null
/mnt/shared/OVERLOAD/work/lang-c/git/abspath.c:40
 	for (start = remaining->buf; is_dir_sep(*start); start++)
...

will not crash as before.