Add stub version.py to appease mypy (#766)
We were using inline ignore to make mypy ignore the missing version.py file
that is generated on install/build but this was causing issues
when the file existed in local mypy runs. This is fixed by adding a stub
version.py file. Since the file is in .gitignore any modifications in
local builds should not make it into version control anyway.