Use latest pfff and fix regression
Thx to Emma we now have more typing information on certain nodes,
but this introduced some regressions because we were comparing
those existing type in the AST of the source to inexistent type in the
AST of the pattern. We should not use the resolved and type information
at the id_info matching level.

Test plan:
make test