Refactor main into OutputHandler
The idea is that all output will be routed through the output handler. This will allow us to collect errors to eventually be added to the JSON, print data as it's ready, and capture output for test purposes.

Not everything is wired up to it yet, I wanted to keep this diff under control.