Add imported entity metavariable as expression in the environment
This fix issue #974
The main issue is that the metavariable matching was working, but
the metavar was added as an ident in the environment, instead of an expr,
and so later one when the same metavariable is used in an expression context,
it was not matching the same metavar that was binded to an ident.

We had similar problems before where a metavar on a function definition
was not matching later the same metavar used in a function call.

Test plan:
test files included