core_runner: use semgrep-core ast cache
Use semgrep-core ast cache while running a set of rules. Automatically
uses a fresh cache (and cleans it up after use) on every run of semgrep
so there is no need to worry about cache invalidation etc.

Change semgrep-core cache argument to take in a path to a directory
to use as a cache. This makes it so that the caller can handle clearing
the cache instead of hard coding a directory that semgrep-core uses.