Added tests for pattern inference (#1167)
* Wrote stub code to test patterns

For a list of file/range pairs, Unit_synthesizer will synthesize the pattern for the given range in the file. Ideally, it will then take that list of patterns, and for each one, check that it matches the code at the range using semgrep.

* Wrote python tests for pattern inference

Added tests that directly match the list with its contents. Filled in python tests based on the file arrays_and_funcs.py. Changed DotAccess to also infer to a metavar.

Ran make test in semgrep-core

* Changed path to relative one

* Separated out the python tests for organizational convenience

* Tests check that each pattern correctly matches the code

When each pattern is iterated through, check that the pattern matches with the expression it was generated with (duplicates some of the Synthesizer code)

Ran with make test (bonus check: before AST was resolved, metrics.send(...) failed to match send(...), and indeed an error was thrown)

* Implemented typed metavar as pattern option when applicable

Wrote java tests for type matching, but did not add them to be ran because of the patterns could not yet parse. Had pattern_from_code also infer types for metavars.

* Added an mli for Synthesizer