Add a new -maxout to limit memory consumption
In the past semgrep-core provoked the reboot of the machine of a few users
because it was using lots of memory (probably because of >||> and the
explosion of paths with '...').
This diff adds the possibility to limit the amount of memory
semgrep-core should use. It then returns an exception instead of crashing
before exhausing all the swap of the user's machine.

Test plan:
uncomment the line in the diff about Gc.full_major and then:

time yy -verbose -lang js -maxout 10_000_000 -rules_file ../../data/basic.yml   l10000.js
+ /home/pad/github/semgrep/semgrep-core/_build/default/bin/Main.exe -verbose -lang js -maxout 10_000_000 -rules_file ../../data/basic.yml l10000.js
maxout allocated memory: 72597504
{
  "matches": [],
  "errors": [
    {
      "check_id": "FatalError",
      "path": "l10000.js",
      "start": { "line": 1, "col": 1 },
      "end": { "line": 1, "col": 1 },
      "extra": { "message": "Fatal Error: Out of memory", "line": "/**" }
    }
  ],
  "stats": { "okfiles": 0, "errorfiles": 1 }
}
0.669 secs