Added test for new equal feature
Currently this test does not pass. See behavior with "make test". Note that the error message is somewhat misleading, since pattern_to_string is used to output the code it should match. In reality, the code it should match is as specified by range and file. In this case, it is

 resp = r.set_cookie("sessionid", resp, "RANDOM-UUID")

(though of course import resolution must be taken into account)