Match filenames in metavars as literal strings (#1310)
Previous work (8cc2954a) added the ability to match metavars to file
names within directives. This work was done specifically to enable
matching and reporting of Go imports.

However, reporting was incomplete here, as pretty-printing has not been
implemented for directives or their tributary AST nodes.

Here I instead bind a literal string expression to the metavariable,
rather than a FileName node. Since we have pretty-printing for such
expressions, reporting now works.

A side effect of this change is that the following pattern:

  import "$X"

  x = $X

will match

  import "fmt"

  x = "fmt"

with $X bound to "fmt", which is a tad odd. This doesn't seem
particularly awful, and is particularly unlikely, so it's left as
a dangling edge case for now.

I've added a test pattern that can be used to verify functionality:

  semgrep --config semgrep-core/tests/go/imports2.sgrep \
    semgrep-core/tests/go/imports.go