Match "OtherAttribute" attributes in any order
We already have code to match keyword and named attributes in any order
in an attribute list, and to match with "less is ok" functionality.

This commit extends that to work with "OtherAttribute" AST nodes as
well. Note that it is not fully implemented here, as I took a shortcut
that would leave this pattern:

```python
@x("a")
@x("a")
def foo:
  ...
```

still matching this code:

```python
@x("a")
def foo:
  pass
```

I'm leaving this as a TODO for now.

Fixes #1406.