Debug logs were being hidden even with --verbose flag
To prevent logs being handled twice semgrep does not propagate logs
from semgrep module to root logging handler so calls to logging.info/debug
were not being handled properly. Really we should not be sending logs
to the root logger and should go through the logger of the submodule.

Fixed calls to go through the correct logger and added a semgrep
rule to alert when the root logger function is used.