Match generic unbraced blocks to single-statement blocks
This fixes a regression from returntocorp/pfff@53808c5690, which
explicitly identified blocks on conditionals.

This previous change was necessary to distinguish between conditionals
with no else block, and conditionals with an empty else block.

However, as a side effect it caused us to no longer match
single-statement blocks with unbraced if blocks.

This commit adds that equivalence in the Semgrep layer.

Fixes #1548.