Use latest ocaml-tree-sitter-lang
Martin fixed the issue with unnamed pattern-tokens
so we have no more Blank and have real tokens!
He also updated the Ruby grammar which was modified quite heavily
which led to quite a few changes in the CST and boilerplate.

test plan:
+ /home/pad/github/semgrep/semgrep-core/_build/default/bin/Main.exe -dump_tree_sitter_cst string.js
|   | Exp_stmt
|   |   | Exp
|   |   |   Assign_exp
|   |   |   |   Choice_member_exp
|   |   |   |   | Id
|   |   |   |   | a
|   |   |   | "="
|   |   |   |   Choice_this
|   |   |   |   | Str
|   |   |   |   |   DQUOT_rep_choice_imm_tok_pat_de5d470_DQUOT
|   |   |   |   |   | "\""
|   |   |   |   |   |   | Imm_tok_pat_de5d470
|   |   |   |   |   |   | foo
|   |   |   |   |   | "\""
|   |   | Auto_semi
|   |   | ""

We now have the content of the string 'foo'! no more Blank