Add a 'none' language to config files to support arbitrary regex scanning. (#1584)
* Add a 'none' language to config files.

'none' language only supports 'pattern-regex', 'patterns', and 'pattern-either' clauses.
The 'none' language makes it easy to support scanning using pattern-regex without having
to specify target files on the CLI.

This also helps address the divergent behavior between includes on the command line and
the 'paths' and 'include' key. The 'paths' and 'include' key did not explicitly scan
included files when in the config file. This behavior persists UNLESS using the 'none'
language, in which case it will scan all files (still respecting the behavior of
all includes and excludes, both on the CLI and the config file.

This commit also added support for HTML comments with '--test'.

Tests are included for this feature.

* Revert __init__.py

* Try to reset pfff/ to develop

* Fix broken span when validating none-language rules

* Make NONE_LANGUAGE the Language type

* Fix issue where I accidentally cut and paste lines in the wrong place -_-

* Update CHANGELOG.md