Fix Parse_tree_sitter_helpers.combine_infos
This will help https://github.com/returntocorp/semgrep/issues/1609

We were not generating the right Parse_info.t (aka token info) in
tree-sitter when we had to combine multiple tokens in one.
Indeed, in pfff we currently parse certain strings as a single
token, but in tree-sitter this was decomposed in 3 tokens, one for the
opening ", one for the content, and one for the closing ".
Before this diff, we were generating the right literal content, but
the token associated with it was just the token for the opening ".
This led to some issues in the editor when a message was mentioning
a metavariable that was bound to a string, because we are using the
tokens to pretty print its content.

test plan:
+ /home/pad/github/semgrep/semgrep-core/_build/default/bin/Main.exe -full_token_info -diff_pfff_tree_sitter string.go
NOTE: consider using -full_token_info to get also diff on tokens

No more diff!