Use junit-xml fork
junit-xml 1.9 does not have a tarball published on pypi. This means
in homebrew formula we have to pin to a git resource which makes it
difficult to autobump the formula. This PR relies on a fork of
junit-xml that has the tarball published.

Relevant ticket for original project to publish tarball has no updates
after 6 months: https://github.com/kyrus/python-junit-xml/issues/69