validate_yaml: Fix crash on 3+ layered jsonschema errors
We were previously able to find where an error came from in a YAML doc
only if a path from the document root
was directly on the error,
or was on the error's parent.

This commit changes the logic so that:
1. We use the absolute path of the error, no the relative one
2. We always use the path of the topmost parent

This changed two existing test cases by not outputting `rules:`
in the error message anymore.
Not exactly sure why this happened,
but I like the change anyway.

Changes were tested by

1. Running the rules from SEMGREP-2P manually
   until we printed a nice error message
   instead of the traceback we had before.
2. Adding a `syntax/bad-three-layers.yaml` test case

Fixes SEMGREP-2P