Add 'metavariable-comparison' frontend to expression eval
Fixes #1724.

Things notably out-of-scope for the first iteration of this functionality:

1.) Multi-language expressions
2.) Multi-metavariable expressions

Per (1), let's start small with a language we know well and expand from
there as needed. Per (2), we can leverage the logical AND functionality
of 'patterns' here instead of multiple metavariables in the same
expression. Again, let's start small and iterate as needed.