Add 'metavariable-comparison' frontend to expression eval (#1824)
* Add 'metavariable-comparison' frontend to expression eval

Fixes #1724.

Things notably out-of-scope for the first iteration of this functionality:

1.) Multi-language expressions
2.) Multi-metavariable expressions

Per (1), let's start small with a language we know well and expand from
there as needed. Per (2), we can leverage the logical AND functionality
of 'patterns' here instead of multiple metavariables in the same
expression. Again, let's start small and iterate as needed.

* Bump CHANGELOG

* Update for latest JSON changes, and add invalid comparison expression handling/test

* Use correct variable

* Bad CHANGELOG overwrite

* Don't fail hard if we encounter a variable name or other string

* Allow base of 0

* Add 'numeric' note