Use metavariable ranges
Before this commit semgrep was using the abstract_context associated
with each metavariable assuming it was what the metavariable matched with
this is not always the case since the abstract context removes some concrete
syntax.

This PR relies on the more stable line ranges reported with the metavariable
as the value of the metavariable.