Do not follow symlinks
Symlinks break semgrep, as they often point outside of the path actually
mounted into the core library, or are otherwise unresolveable by the
core.

In the tradition of comparable UNIX tools (such as ripgrep), we disable
link following.

In the future, we may want to add a switch to allow link following, but
let's wait until this is actually needed to avoid flag bloat in the
immediate.

Fixes #2005.