Parse using block in C#
E.g.
```csharp
using (Font font3 = new Font("Arial", 10.0f), font4 = new Font("Arial", 10.0f))
{
   // Use font3 and font4.
}
```

We convert the using block to a `try` block, with the declaration outside the
try. This is also what C# does internally. Should we call `Dispose` in the
finally? Can we know on what we should call `Dispose`? Is this the correct way
forward, or should we use `OSWS_With`?