Refactor matching to use Metavars_generic.Id not Metavars_generic.E
Closes https://github.com/returntocorp/semgrep/issues/2156

We used to do some hacky things when a metavariable was matching
an identifier such as an id in an expression, or class name, or type.
If the user reuse the same metavariable in different context, we
want the metavariable to match all those contexts if the resulting
is is equal. For example class $X { ... $X foo() { ... } } where
$X is here an entity name (in AST_generic.entity.name), or
a return type (in AST_generic.type_.TyId).

In a expression context, an id has also an id_info with inside
a single unique id assigned during the naming phase. This allows
us to make sure multiple use of the same metavar refer to the same
variable (this is especially useful with deep statement matching where
we can have lots of variables using the same name but in very different scope).

With this diff we add the binded metavariable as an Id with an optional
id_info attached to it, instead of adding it as an (E (Id ...)) which
was then requiring some hacky things.

Test plan:
test file included
make test