Allow 'self.$FOO = $FOO' to match self.bar = bar
I'm not sure we should allow this, because the $FOO correspond
to very different entity (a field and a param), and it just happens
that they use the same name.
Anyway, some users seems to intuitively want the pattern to match
so let's try it.
At least we don't allow variable in different scopes to be matched against
each other even if their names are equal.

This closes https://github.com/returntocorp/semgrep/issues/2304

test plan:
test files included.
make test