Fix bug that caused m_compatible_type to only bind the type
Previously, when matching a type with an id, we would only call m_type_option_with_hook, which would add the metavariable for the type into the environment if applicable, but not the actual typed metavar. Similarly, the C string metavariable pointer type change made the same error.

Now, we bind the matcher to the matcher for the typed metavar.

Test plan: make test with cases in metavar_typed_reuse