Use bloom filters to reduce tree traversal
Closes #2352

When deciding whether or not to match a subtree, we can definitely choose not to if essential parts of the pattern are not present. It is cheap and space efficient to first hash these essential parts using a bloom filter, which will guarantee a "no" answer if it is not present. Then, when matching the pattern against a node, we only traverse further if all essential parts of the pattern are present in that node.

For the current implementation:
- the essential parts are non-metavariable idents and string literals
- a bloom filter is stored at every statement node

The initial bloom annotation is worst case O(sn), where s is the number of strings and n is the number of statement nodes. This is because each string for a given statement is added to that statement's bloom filter and each of its parents'.

Later computations will be O(s_p), where s_p is the number of strings in the pattern.

Test plan: make test (correctness should still hold), benchmarking to see if it is faster. Will also need to tweak the parameters passed into Bloomf.create

This implementation makes use of the bloomf library in opam