Use bloom filters to reduce tree traversal (#2464)
* Use bloom filters to reduce tree traversal

Closes #2352

When deciding whether or not to match a subtree, we can definitely choose not to if essential parts of the pattern are not present. It is cheap and space efficient to first hash these essential parts using a bloom filter, which will guarantee a "no" answer if it is not present. Then, when matching the pattern against a node, we only traverse further if all essential parts of the pattern are present in that node.

For the current implementation:
- the essential parts are non-metavariable idents and string literals
- a bloom filter is stored at every statement node

The initial bloom annotation is worst case O(sn), where s is the number of strings and n is the number of statement nodes. This is because each string for a given statement is added to that statement's bloom filter and each of its parents'.

Later computations will be O(s_p), where s_p is the number of strings in the pattern.

Test plan: make test (correctness should still hold), benchmarking to see if it is faster. Will also need to tweak the parameters passed into Bloomf.create

This implementation makes use of the bloomf library in opam

* Create a list of strings in the pattern rather than a bloom filter for them

* Renamed in semgrep_generic and fixed infinite loop

Changed bf to pattern_strs where it originally represented the bloom filter

Had infinite loop because the visitor visits the top level statement. Hacky fix--just always skip the first statement reached

* Ignore string literal semgrep constructs (..., regexp)

* Traverse id_info in visitor

Added a hook for id_info. To maintain the same default behavior, v_id_info enters the refs, but the default hook kid_info does not. Bloom_annotation overrides that by repalacing the hook with one that traverses

* Used constant propagation information and added ident special cases

Add hook to constness to match literals. For idents, it was necessary to exclude strings starting with "$..." and regexps. This is currently done poorly.

* Removed print statements

* Resolvd conflict

* Also filter pattern in generic_vs_generic

Constant propagation test case in java has more cases so that I could measure speedup

* Reverted Generic_vs_generic changes

Use of bloom filter check in generic_vs_generic did not make a difference

* Returned must_analyze_statement_bloom... func to semgrep_generic

* Removed print statements

* Addressed comments

* Use Str.first_chars str 4 instead of String.sub

TODO: use a regex

* Profile bloom annotating

* Only check for metavariables/ellipses in the pattern

* Added -bloom_filter to turn on the bloom filter, default off

* Added a comment, removed a print statement