Start integrating spacegrep directly in semgrep-full-rule-in-ocaml
We want semgrep-core to be able to use full rules (not just "mini rules")
so we need to handle the [generic] language and calling spacegrep
directly.

This is the first part towards spacegrep integration in semgrep-core.
We just parse the pattern; we currently do not run the engine.

test plan:
make test