matching: Match variables with literals modulo constness
As a result of adding this feature, we have to be more rigorous when
assigning constness to variables. Now setting constness to a literal
means that the specific occurrence of the variable can be replaced by
the literal. And in `x = E` the variable `x` stands for a label rather
than a value, so we cannot set its constness to any other value than
_Top_ (`NotCst`).

Closes #374

test plan:
make test