Use the yaml package to parse yaml code into the generic AST (#2630)
The yaml package we use to parse ocaml has a low level stream API that gives the tokens lexed. We use this stream API to turn yaml into a generic AST, with manually generated tokens. These tokens are all FakeTokStr, but the position information is set. Note that if you have

- id: some_id

this will consider the column of "id" to be 2. Values are 0 indexed (this may need to be changed?)

Does not currently handle alias/anchors or semgrep patterns. Doing this may require manually separating out ellipses

Test plan: parsing the below yaml snippet should result in
rules:
- id: my_pattern_id
  pattern: |
    if ( $X ) $Y;
  message: |
    Semgrep found a match
  severity: WARNING

E(
  Container(Dict,
    [Tuple(
       [L(String(("rules", ())));
        Container(Array,
          [Container(Dict,
             [Tuple([L(String(("id", ()))); L(String(("my_pattern_id", ())))]);
              Tuple(
                [L(String(("pattern", ())));
                 L(String(("if ( $X ) $Y;
", ())))]);
              Tuple(
                [L(String(("message", ())));
                 L(String(("Semgrep found a match
", ())))]);
              Tuple([L(String(("severity", ()))); L(String(("WARNING", ())))])])])])]))