Remove the ... optimisation which introduces some regressions
This helps https://github.com/returntocorp/semgrep/issues/2646
but this is not the right fix.
The only difference is that we return multiple times the same thing
in semgrep-core, which should not matter, but it seems to matter
for the semgrep wrapper.
Note that semgrep-core -config is not affected by the PR.

test plan:
$ semgrep --debug --verbose --config
tests/OTHER/rules/regression_uniq_or_ellipsis.yaml tests/OTHER/rules/

returns only 1 match (like in 0.40)