[WIP] Filtering irrelevant rules using regexps by analyzing the whole rule
This is similar to the filtering of irrelevant patterns I was doing
in Mini_rules.filter.ml, but analyzing the whole rule, not individual
patterns to skip entire rules. By also making the generic AST parsing
lazy, we can even sometimes skip parsing a file.

Right now the regexp is hardcoded for the rule used to match
big-js, but subsequents commits will fix that.

test plan:
$ time yy -profile -debug -lang js -config bench/big-js/semgrep_slow.yml
bench/big-js/
=> 5m27
$ time yy -profile -filter_irrelevant_rules -debug -lang js -config bench/big-js/semgrep_slow.yml bench/big-js/
=> 0.6s