Add profiling for method causing slowness in python rule
Added the rule and profiling on Stmts_match_span.location

This rule is slow because we need to get the location of many matches spanning the entire tree.

We can solve it in two ways. One, we can predetermine the range of each statement (easy to add), as well as each statement list, so that work is not being repeated. Two, now that we analyze the whole rule, we should not run pattern-not on the entire rule, but only the considered matches.

Test plan:

semgrep-core -profile -bloom_filter -config tests/OTHER/perf1/benchmarking/netflix-slow-python.yml ../perf/bench/netflix/input/lemur -lang python