Make deep expressions traverse OtherStmt, Lambda nodes
A few easily recurseable cases were omitted from deep-statement
matching: OtherStmt nodes, and Lambda nodes. The former can refer to an
arbitrary number of `any` nodes, and we update to traverse any wrapped
`E` nodes. The latter we treat like any other `FunctionDef` node.

The `OtherStmt` change should be relatively innocuous, as these
statements are relatively rare in code. I'm a bit worried about the
performance impacts of the Lambda change, however, as a minified js
file could easily contain a very large number of anonymous functions.

To that end, let's not merge this until we can confirm that we are not
creating a performance regression with this change.

Fixes #2879.