Run performance benchmarks on commit
Restructure run-benchmarks so that it can be used in the test suite that runs on commit. In particular, we only want to run small tests, since these run every commit, so a flag (`small-only`) was necessary to specify that. Also, to reuse compare_perf.py (in scripts), added a flag (`save-to`) to specify a timings file.

Test plan: commit