* core_runner.py: do not call semgrep-core if there are no target
This fixes I think the last regression of --experimental compared
to the std benchmark. --experimental was slower than std because
for many benchs, we use r2c-ci.yml which contain rules that do
not have any target in the repository (e.g., python rules on a js
repo) and we should not lose time calling semgrep-core for those.

test plan:
./run-benchmarks now shows --experimental always faster (minor small
variations)