Run performance benchmarks on commit (#2935)
* Run performance benchmarks on commit

Restructure run-benchmarks so that it can be used in the test suite that runs on commit. In particular, we only want to run small tests, since these run every commit, so a flag (`small-only`) was necessary to specify that. Also, to reuse compare_perf.py (in scripts), added a flag (`save-to`) to specify a timings file.

Test plan: commit

* fix run-benchmarks args

* Fixed path to run-benchmarks

* debugging info

* use correct name of script

* Move matplotlib import to only demand it when necessary

* Fixed flags

* Output semgrep and semgrep-core verisons for debugging

* Disable build-core requirement

* print stderr, stdout for debugging

* Use later version of semgrep

* Don't call semgrep-core in 0.40.0 benchmark

* Remove debugging version print

* Restore build-core dependency

* Wrong folder for timing.txt

* Use 0.46.0 instead of 0.40.0

* Removed accidentally committed files