Use org fork of homebrew-core when creating version bump PR
Before this commit, we were using a private fork to make a PR
to homebrew-core to bump the version of semgrep. After brew 3.1.3
it is possible to add a --fork-org flag to specify using an org
fork for bump-version-pr. This PR makes use of this new flag in the
bump-version-pr github action.