Use a canonical string for each language
Define a language enum and map user given strings (eg "csharp", "cs", "C#") to the appropriate language enum, which has a specific value. This allows the list of languages to be treated as a set, so that there will never be duplicate languages.

This fixes a bug where `--optimizations` returned typescript results multiple times