matching: Prefer Id over E when binding typed metavariables to names
When a metavar binds to a name this must be represented using the Id
constructor.

This fixes a -config regression wrt Python's Semgrep wrapper affecting:
semgrep-rules/java/lang/security/audit/xss/no-direct-response-writer.java

The problem can be reproduced with a simple pattern-and:

    patterns:
    - pattern: |
        (PrintWriter $WRITER).println(...)
    - pattern-not: $WRITER.println("...")

on this Java code:

    PrintWriter writer = response.getWriter();
    writer.println("foo");

The typed metavar $WRITER used in the positive pattern is bound to
`MV.E (N (Id "writer"))` whereas the non-typed metavar $WRITER used in
the negative pattern is bound to `MV.Id "writer"`. Then, when we check
if there is a negative match that filters out our positive match, we
can't find one because `MV.E` and `MV.Id` are different kinds of
values.

test plan:
make test # test included