Add special case for whole rule pattern parse errors (#3277)
With whole rule running, semgrep can no longer use the
pattern -> character numbers mapping that it uses to pretty print
pattern parse errors. We currently do not have access to the correct
character offsets in the yaml parser in semgrep-core but are thinking
of moving to a differnt parser that does give us character offsets.

For now special case the pattern-parse errors when running with
whole rule running optimization.