Using "var" when an explicit type is expected is an error (#3281)
The parser sees "var" as just any type, but it can't be used in many places
where a specific type is required. We disallow it in `type_`, and only allow it
in `local_variable_type`.

This was already more or less the case, but now it's explicit and explained in
a comment.