ci: wait 10 min after pypi to publish to homebrew (#3303)
brew bump-formula-pr depends on pipgrip to update dependencies
and pipgrip seems to not work right after publishing to pypi but seems
to work fine when run a few minutes later. Suspect it is pypi propagating
the latest release. Add a CI job that waits 10 minutes before trying to
publish to homebrew.