Use logger level as single source of truth instead of verbosity globals
Before this commit we were using DEBUG and QUIET globals to decide when
things should or should not be logged (we cannot rely on just the logging
framework's handlers because tqdm uses prints). This PR uses whatever logging
level is stored by the root logger as the source of truth instead of relying
on the globals.