Allow users to disable certain matching features in a rule
This will allow to disable for example constant propagation
on a per-rule basis.

I've chosen the 'settings' keyword, but we could change it
for 'preferences', or 'config', whatever.

Still need to update the jsonschema after so it can be used
from the semgrep python wrapper.

test plan:
test file included