fix: use --debug flag to gate semgrep-core debug output instead of --debugging-json
Prior to this commit, semgrep-core output was only being printed out when semgrep
was called with `--debugging-json` but not when `--debug` was set. This PR makes
the output of debug info more consistent with the rest of our codebase.

Now `--debugging-json` is only used to add step-by-step pattern match evaluation
logs. Holding off on renaming for now since the OutputSettings class is potentially
used externally.