Refactor Match_rules somewhat (#3416)
* Refactor Match_rules somewhat

Factorize the repeated code in `check` and what was `evaluate_nested_formula`. This function takes a formula and produces the matching ranges, so we call it `evaluate_formula`. The old `evaluate_formula` is now renamed to `compose_patterns`.

Also, rename the methods that return bools to make it more obvious. These names might not be ideal for the functionality.

Test plan: this should change no functionality

* Remove annotation

* Remove profiling in mutually recursive function

* Iago's comments