core: Avoid crashing due to fake tokens
Hopefully one day we can guarantee that every expression has a known
range. But, in the meantime, if we cannot get a range, we should handle
the situation more gracefully.

Let's remove Visitor_AST.range_of_any in favor of the safer
Visitor_AST.range_of_any_opt. This will force us to think what to do
in each situation where range info is not available. For now we just
log an error and move to something else.

Follows: #3603

test plan:
make test