Do something about stack overflows (#3625)
* Print a warning when the stack size exceeds 7 MB.
Raise a 'Stack_overflow' exception when the stack size exceeds half of
the user-specified memory limit (requires 'ulimit -s unlimited' or
equivalent).

* Document the issues with stack size limits

* Add '-max_stack' option to semgrep-core

* Add missing files

* Add tests for memory limits

* Add missing dune file

* Try more generous numbers for stack overflow tests

* Only call 'Gc.compact ()' as a desperate move when catching an
'Out_of_memory' exception because it's an expensive operation.
Calling it every time was making semgrep 6x slower.

* Update changelog

* Remove -max_stack option because it's not reliable and only useful on
some platforms. The stack warning threshold however is lowered to 100
KiB. It is also not reliable but it's only a warning that hopefully
will be noticed now and then.