Many fixes to improve the parsing_stat CI job (#3664)
* Many fixes to improve the parsing_stat CI job

The Lua parsing stat job was failing because it was using
a non-github repo which triggered an RPC error at clone time.

The Kotlin parsing stat job was failing because it was using
CST.dump_tree which output things on stdout, which then messes up
the json.

The C parsing stat job was failing because of some out of memory
fatal errors on huge files. I tried to use Memory_limit
but ran in many issues. It's hard to intercept those Out_of_memory
reliably.

test plan:
./run-lang --upload lua
./run-lang --upload kotlin
./run-lang --upload c

* iago's comment