Do not match "..." to interpolated strings
This is a quick fix to immediately improve Semgrep's consistency, but it is definitely not a good one and we should change our implementation of interpolated strings.

The problem was that `x = "..."` does not match `x = "str{input}"`, but `"..."` matches `"str{input"`. This is especially unfortunate because rule writers will use `"..."` to search for constant strings, and there is no other syntactic way (that I could figure out) to do this.

TODO: A better solution might be to create an InterpolatedLiteral expression and have `"str"` in the above example be an `InterpolatedLiteral(String("str", _))` instead of `L(String(str, _))`. This is fairly time consuming and should be thought through more carefully first, though, and I wanted to be able to write my rule.