[Java] fix the range and autofix of Cast expression
This closes https://github.com/returntocorp/semgrep/issues/3669
This is a bit ugly. The right fix would be to extend
the AST_generic.expr record to store the leftmost and rightmost
token in the expr, so we would not need to abuse tok in the
expr_kind itself.

test plan:
test file included