Remove legacy metavariable interpolation code
We had to keep track of all metavariables because we used to run
semgrep-core over single patterns instead of over entire rules. This
meant that some patterns did not have the metavariable a rule was using.

Now that we run over the entire rule, match object returned by
semgrep-core should have all the metavariables needed for interpolation.