use_parsing_cache: Handle SIGXFSZ (file size limit exceeded) (#3843)
Thanks to Martin and Yoann for valuable feedback.

NOTE: Initially I thought that we were creating huge files in some cases. It turns out it is not the case. But, somehow, something is setting a small file size limit in some of my terminals triggering this problem. In any case, the changes are minimal, and I believe it's still a good idea to be defensive, catch errors during writing, and clean up the cache files to avoid corruption.

Closes PA-318

test plan:

    % semgrep-core -lang js -j 1 -e '$X == $X' -fast \
          -use_parsing_cache ./cache \
          ~/semgrep/perf/bench/juice-shop/input/juice-shop/
      #^ no longer gets killed by SIGXFSZ (Apple M1, MacOS 11)
      # and with -j 2 we no longer get a segfault