Add support for `--config auto`
This will support rule recommendations from the CLI.

Note that this won't be functional until a corresponding PR on
semgrep.dev is merged (#3554).

In order to serve rule recommendations for a project, we send the
project URL on config requests.

I turned off the spurious `disallow_any_decorated` rule in mypy,
as it was just causing us to sprinkle type ignore comments around our
codebase, thus hiding _actual_ type problems.