optimizing: Analyze_rule: Fix CNF simplification in step1
When a term in an OR cannot be checked with a regex, we have to assume
that the term holds and thus reduce the entire OR to "true".

Plus, enable -filter_irrelevant_rules for full-rule tests.

Plus, prevent Analyze_rule from crashing when given a "new" formula
where things like `Not (Or ...)` are possible, otherwise e.g. test
tests/OTHER/rules/negation_ajin fails (by `failwith "Not Or"`).

Closes #3755

test plan:
make test # tests included

TL;DR

For example:

    - pattern-either:
        - patterns:
            - pattern: $X.sha1("$RE")
            - metavariable-regex:
                metavariable: $RE
                regex: (?i)md2
        - patterns:
            - pattern: $X.getMd2Digest(...)

This is seen by Analyze_rule as:

    OR
        AND (contains "sha1") (dont-know-how-to-check)
        AND (contains "getMd2Digest")

Sometimes Analyze_rule does not know how to extract a regex from a
pattern, here we represent that with `(dont-know-how-to-check)`. The
"problem" with the `metavariable-regex` here is that Analyze_rule does
not know how to handle the case-insensitive mode `(?i)`.

When converting the regex formula into CNF we get:

    AND
        OR (contains "sha1") (contains "getMd2Digest")
        OR (dont-know-how-to-check) (contains "getMd2Digest")

Then in step1 we try to get rid of these `(dont-know-how-to-check)`. The
bug was caused by incorrectly simplifying this as:

    AND
        OR (contains "sha1") (contains "getMd2Digest")
        OR (contains "getMd2Digest")

That is, treating `(dont-know-how-to-check)` as "false" rather than
"true". The correct simplification is:

    AND
        OR (contains "sha1") (contains "getMd2Digest")