Match deep ellipsis against statement sub-expressions
Previously, we only supported deep-ellipsis for matching against
expression sub-expressions.

That is, the following pattern:

  if (cond) {
    <... a ...>
  }

will match:

  if (cond) {
    some_complicated_expression_of(a);
  }

However, we would not match statement sub-expressions. So the pattern:

  <... a ...>
  b

would not match:

  if (a) { ... }
  b

This commit generalizes m_expr_deep to enable use with either expression
or statement sub-expressions (actually to anything for which we can
extract sub-expressions), and then adds a matching branch for
deep-ellipsis against statements.

Closes PA-41.