tainting: Introduce a special kind of sanitizer (#4033)
In PA-402 we made all sanitizers "not-conflicting" so that one could
declare `$F(...)` as a sanitizer without conflicting with sources and
sinks. That makes sense in that particular use case, but it partially
breaks our ability to simulate sanitizers by side effect. For example:

    pattern-sanitizers:
      - patterns:
          - pattern-inside: |
              $JWT.verify($TOKEN, ...)
              ...
          - pattern: $TOKEN
    pattern-sinks:
      - patterns:
          - pattern: $JWT.decode($TOKEN, ...)
          - pattern: $TOKEN
    pattern-sources:
      - pattern: $TOKEN

no longer works on this example:

    jwt.verify(token, key);
    // ok: test
    jwt.decode(token, true);

Since the second `token` is both matched by a sanitizer and a sink (and
a source), the sanitizer is now "disabled" due to the changes made by
0b35ab5004c in connection with PA-402. This is not at all desirable, but
we still want the ability to declare `$F(...)` as a sanitizer. Then the
simplest solution seems to be to support two different kinds of
sanitizers.

Closes: PA-455
Changes: 0b35ab5004c ("tainting: Filter out sanitizers that conflict with sources/sinks (#3958)")

test plan:
make test # tests included