constness: Track constness of names and nothing else
When translating `x.f(y)` to IL we did not carry the constness of `x`,
so even if `x` was a constant at that point, we would not recognize it
as such in that specific context.

Initially we tracked constness in the `lval`s to later support field
sensitivity, but this is not yet implemented and just adds confusion.
For now it is pointless given that we only track the constness of
simple variables.

By making this simpler, the `x.f(y)` bug almost fixes itself.

Closes PA-588

test plan:
make test # test included