analyzing: cfg: Fix `try` blocks ending in `return`
When building the CFG of a `try-catch` we connect the end of the
try-block to the beginning of the catch-block. But, if the CFG of the
try-block ended in `return`, this did not happen, and the catch-block
became dead code, despite something in the try-block could throw an
exception.

Closes #4091

test plan:
make test # tests included