analyzing: CFG_build: Make sure labels are resolved
We did not create a dummy node to apply pending labels to when the
preceeding statement exited the function. We must create the dummy node
and label it, otherwise some `goto` may be left unresolved.

This also fixes the "Could not resolve label: _label" warnings that we
had been observing for a while.

Fixes: 20cd8a784b8 ("Add support for Break and Continue in IL (#3925)")

Closes #4265

test plan:
make test # test included