logging: Reduce the size of --debug dumps
In a first step, we typically just want `semgrep --debug` to help us
identify which files are causing problems. For this we use the default
INFO level. If a message is logged rarely then it does no harm having it
at the INFO level. But, if it is logged often, then it may just be making
the dump larger without really helping that much in most situations.

This will move some frequently-logged messages to the TRACE level to
reduce the size of --debug dump. My intuition is that these messages are
not particularly useful when one is trying to find "the file" causing
trouble.

test plan:
$ semgrep --time --debug -c p/auto lang/javascript/tmp/* &>log
The size of the log files went down from 838M to 174M (~5x smaller).